feat: finish timer logic
